# Backend/OpenAPI/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApi",
    "version": "v1"
  },
  "paths": {
    "/api/bannedWords": {
      "get": {
        "tags": [
          "BannedWords"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "BannedWords"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "BannedWords"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/bannedWords/{id}": {
      "delete": {
        "tags": [
          "BannedWords"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/files/uploadFile": {
      "post": {
        "tags": [
          "File"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": [
          "File"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/files/{id}": {
      "delete": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": { }
}